library/views.py
from rest_framework.generics import ListAPIView

from library.models import Book, Category
from library.serializers import BookSerializer, CategorySerializer


class CategoryListAPIView(ListAPIView):
    serializer_class = CategorySerializer
    queryset = Category.objects.all()


class BookListAPIView(ListAPIView):
    serializer_class = BookSerializer
    queryset = Book.objects.all()
